get overlapping svs

+++ helper.py
@@ -763,28 +763,10 @@ def compare_short_long_reads():
     merged.to_csv("1kgp/sr_lr_merged.csv", index=False)
 
 
-def get_overlapping_clustered_svs():
-    """Checks for overlaps between clustered SVs and the original SV breakpoints."""
-    lookup = get_sv_lookup()
-    lookup["sv_id"] = lookup["id"].astype(str)
-    lookup.rename(columns={"stop": "end"}, inplace=True)
-    bed_file = "1kgp/sv_lookup.bed"
-    df_to_bed(in_df=lookup, out_file=bed_file)
-
-    output_bed_file = "1kgp/sv_lookup_intersect.bed"
-    subprocess.run(
-        ["bash", "bed_intersect.sh"]
-        + [
-            bed_file,
-            "1kgp/consensus_svs.bed",
-            output_bed_file,
-        ],
-        capture_output=True,
-        text=True,
-    )
-
+def bed_to_df(bed_file: str, remove_dupes: bool = False) -> pd.DataFrame:
+    """Convert a BED file to a pandas DataFrame."""
     df = pd.read_csv(
-        output_bed_file,
+        bed_file,
         delimiter="\t",
         names=[
             "sv1_chr",
@@ -799,19 +781,70 @@ def get_overlapping_clustered_svs():
             "sv2_svid",
         ],
     )
+    if remove_dupes:
+        df = df[df["sv1_id"] != df["sv2_id"]]
 
-    # check for 25% reciprocal overlap between svs
     df["r"] = df.apply(
         lambda row: reciprocal_overlap(
-            (row.sv1_l, row["sv1_r"]), (row["sv2_l"], row["sv2_r"])
+            (row["sv1_l"], row["sv1_r"]), (row["sv2_l"], row["sv2_r"])
         ),
         axis=1,
     )
+    return df
+
+
+def get_overlapping_clustered_svs():
+    """Checks for overlaps between clustered SVs and the original SV breakpoints."""
+
+    if not os.path.exists("1kgp/sv_lookup_intersect.csv"):
+        lookup = get_sv_lookup()
+        lookup["sv_id"] = lookup["id"].astype(str)
+        lookup.rename(columns={"stop": "end"}, inplace=True)
+        bed_file = "1kgp/sv_lookup.bed"
+        df_to_bed(in_df=lookup, out_file=bed_file)
+
+        output_bed_file = "1kgp/sv_lookup_intersect.bed"
+        subprocess.run(
+            ["bash", "bed_intersect.sh"]
+            + [
+                bed_file,
+                "1kgp/consensus_svs.bed",
+                output_bed_file,
+            ],
+            capture_output=True,
+            text=True,
+        )
+
+        df = bed_to_df(output_bed_file)
+        df.to_csv("1kgp/sv_lookup_intersect.csv", index=False)
+    else:
+        df = pd.read_csv("1kgp/sv_lookup_intersect.csv")
 
     overlaps = df[(df["r"] >= 0.25) & (df["sv1_id"] != df["sv2_id"])]
-    print(overlaps[["sv1_id", "sv2_id", "r"]])
 
-    df.to_csv("1kgp/sv_lookup_intersect.csv", index=False)
+    # this gets all overlapping SVs - however, we want to check whether the original SVs were overlapping as well (pre-clustering)
+    # print(
+    #     overlaps[["sv1_id", "sv1_l", "sv1_r", "sv2_id", "sv2_l", "sv2_r", "r"]]
+    # )
+
+    og_overlaps = pd.read_csv("1kgp/og_svs_intersect.csv")
+    og_overlaps = og_overlaps[og_overlaps["r"] >= 0.25]
+    # only need sv ids to check for overlaps
+    og_overlaps = og_overlaps[["sv1_id", "sv2_id"]]
+
+    # merge new overlaps with og_overlaps
+    new_overlaps = overlaps.merge(
+        og_overlaps,
+        left_on=["sv1_id", "sv2_id"],
+        right_on=["sv1_id", "sv2_id"],
+        how="left",
+        indicator=True,
+    )
+    # filter for entries in overlaps that are not in og_overlaps
+    new_overlaps = new_overlaps[new_overlaps["_merge"] == "left_only"]
+
+    # TODO: sv2 coordinates do not take into account the read length (i.e. SV is 450 bp longer than reference)
+    print(new_overlaps)
 
 
 def write_post_processed_files(stem: str = "1kgp"):
